requirements.txt
alabaster==0.7.9
appdirs==1.4.0
argh==0.26.2
Babel==2.3.4
cffi==1.9.1
cryptography==1.7.2
docutils==0.13.1
ezhost
Fabric3==1.13.1.post1
idna==2.2
imagesize==0.7.1
Jinja2==2.9.5
livereload==2.5.1
MarkupSafe==0.23
packaging==16.8
paramiko==2.1.1
pathtools==0.1.2
port-for==0.3.1
pyasn1==0.1.9
pycparser==2.17
Pygments==2.2.0
pyparsing==2.1.10
pytz==2016.10
PyYAML==3.12
requests==2.13.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.2
sphinx-autobuild==0.6.0
tornado==4.4.2
watchdog==0.8.3
